+++ webpack.common.js
@@ -1,4 +1,5 @@
 const path = require('path')
+const Dotenv = require('dotenv-webpack')
 
 module.exports = {
   entry: {
@@ -49,5 +50,10 @@ module.exports = {
         loader: 'url-loader'
       }
     ]
-  }
+  },
+  plugins: [
+    new Dotenv({
+      systemvars: true
+    })
+  ]
 }