+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [5.6.4](https://github.com/digidem/mapeo-desktop/compare/v5.6.3...v5.6.4) (2025-03-05)
+
+
+### Bug Fixes
+
+* strict identification of ISO dates ([#758](https://github.com/digidem/mapeo-desktop/issues/758)) ([dc07b9d](https://github.com/digidem/mapeo-desktop/commit/dc07b9d))
+
+
+
+### [5.6.3](https://github.com/digidem/mapeo-desktop/compare/v5.6.2...v5.6.3) (2024-05-14)
+
+
+### Bug Fixes
+
+* update mapbox token ([b484674](https://github.com/digidem/mapeo-desktop/commit/b484674))
+
+
+
+### [5.6.2](https://github.com/digidem/mapeo-desktop/compare/v5.6.1...v5.6.2) (2023-08-31)
+
+
+
+### [5.6.1](https://github.com/digidem/mapeo-desktop/compare/v5.6.0...v5.6.1) (2023-08-31)
+
+
+
 ## [5.6.0](https://github.com/digidem/mapeo-desktop/compare/v5.6.0-beta.2...v5.6.0) (2022-12-13)